pyproject.toml
[project]
name = "crmconnector"
version = "0.1.0"
description = "CRM Courses Connector"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp==1.18.0",
    "agent-framework",
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-exporter-otlp",
    "opentelemetry-instrumentation-requests",
    "opentelemetry-instrumentation-aiohttp-client",
]

[project.optional-dependencies]
dev = ["debugpy==1.8.8"]